refactor: move base interface to `povm_toolbox.quantum_info.base`
