+++ docs/apidocs/povm_toolbox.quantum_info.base.rst
@@ -1,0 +1,13 @@
+==========
+Interfaces
+==========
+
+.. warning::
+   This API reference is still `under construction`.
+
+.. _povm_toolbox-quantum_info-base:
+
+.. automodule:: povm_toolbox.quantum_info.base
+   :members:
+   :no-inherited-members:
+   :no-special-members:
