+++ povm_toolbox/quantum_info/base/__init__.py
@@ -1,0 +1,22 @@
+# (C) Copyright IBM 2024.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
+
+"""The base interfaces for the quantum-informational frame, POVM, and Dual classes."""
+
+from .base_dual import BaseDual
+from .base_frame import BaseFrame, LabelT
+from .base_povm import BasePOVM
+
+__all__ = [
+    "LabelT",
+    "BaseFrame",
+    "BasePOVM",
+    "BaseDual",
+]
